Notification mechanism changed with ServiceIntent fired by bc-receiver and main activity (reboot case still doesn't work). Woking Products Page with the possibility to modify products icons and other stuff. bye.

+++ app/src/main/java/com/example/smartpantry/ProductPantryItem.java
@@ -7,9 +7,9 @@ public class ProductPantryItem {
     String expire_date;
     String icon;
     boolean is_favorite;
-    int quantity;
+    long quantity;
     ProductPantryItem(String name, String description, String expire_date, String id, String icon,
-                      int is_favorite, int quantity) {
+                      int is_favorite, long quantity) {
         this.name = name;
         this.expire_date = expire_date;
         this.id = id;
